# estudio_madame/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Estúdio Madame API
  description: API para gerenciamento de galerias fotográficas com integração Google Drive
  version: 1.0.0
  contact:
    name: API Support
    email: support@estudiomadame.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.estudiomadame.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token para autenticação
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        avatar:
          type: string
          format: uri
          nullable: true
        galleries_count:
          type: integer
        last_activity:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    GallerySettings:
      type: object
      properties:
        privacy:
          type: string
          enum: [public, private, password]
        password:
          type: string
          nullable: true
        allow_downloads:
          type: boolean
        mature_content:
          type: boolean
        max_client_selection:
          type: integer
          nullable: true
        layout:
          type: string
          enum: [grid, masonry, list]
        default_sort:
          type: string
          enum: [newest, oldest, az]
    GalleryStats:
      type: object
      properties:
        views:
          type: integer
        downloads:
          type: integer
        selections:
          type: integer
    Gallery:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        client_id:
          type: string
          format: uuid
        client_name:
          type: string
        status:
          type: string
          enum: [draft, published, client_selection, archived]
        cover_image:
          type: string
          format: uri
          nullable: true
        photo_count:
          type: integer
        date_created:
          type: string
          format: date-time
        event_date:
          type: string
          format: date
          nullable: true
        location:
          type: string
          nullable: true
        access_token:
          type: string
          nullable: true
        google_drive_folder_id:
          type: string
          nullable: true
        google_drive_folder_name:
          type: string
          nullable: true
        auto_sync_enabled:
          type: boolean
        last_sync_at:
          type: string
          format: date-time
          nullable: true
        sync_status:
          type: string
          enum: [idle, syncing, error]
        settings:
          $ref: '#/components/schemas/GallerySettings'
        stats:
          $ref: '#/components/schemas/GalleryStats'
    CreateGalleryRequest:
      type: object
      required:
        - title
        - client_id
      properties:
        title:
          type: string
        description:
          type: string
        client_id:
          type: string
          format: uuid
        event_date:
          type: string
          format: date
        location:
          type: string
        cover_image:
          type: string
        google_drive_folder_id:
          type: string
        auto_sync_enabled:
          type: boolean
          default: false
        settings:
          $ref: '#/components/schemas/GallerySettings'
    UpdateGalleryRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, published, client_selection, archived]
        event_date:
          type: string
          format: date
        location:
          type: string
        cover_image:
          type: string
        auto_sync_enabled:
          type: boolean
        settings:
          $ref: '#/components/schemas/GallerySettings'

    # ==================== PHOTO SCHEMAS ====================
...
    SyncJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        gallery_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        photos_added:
          type: integer
        photos_removed:
          type: integer
        error_message:
          type: string
          nullable: true

    # ==================== GOOGLE DRIVE SCHEMAS ====================
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
paths:
...
  /galleries:
    get:
      tags:
        - Galleries
      summary: Listar todas as galerias
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, client_selection, archived]
        - name: client_id
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          description: Buscar por title ou client_name
          schema:
            type: string
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [date_created, title, client_name, photo_count]
            default: date_created
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Lista de galerias
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Gallery'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags:
        - Galleries
      summary: Criar nova galeria
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGalleryRequest'
      responses:
        '201':
          description: Galeria criada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Gallery'
        '400':
          description: Dados inválidos
        '404':
          description: Cliente não encontrado
  /galleries/{id}:
    get:
      tags:
        - Galleries
      summary: Obter detalhes de uma galeria
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhes da galeria
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Gallery'
                      - type: object
                        properties:
                          client:
                            $ref: '#/components/schemas/Client'
        '404':
          description: Galeria não encontrada

    patch:
      tags:
        - Galleries
      summary: Atualizar galeria
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGalleryRequest'
      responses:
        '200':
          description: Galeria atualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Gallery'
        '404':
          description: Galeria não encontrada

    delete:
      tags:
        - Galleries
      summary: Deletar galeria
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Galeria deletada com sucesso
        '404':
          description: Galeria não encontrada
  /galleries/bulk-action:
    post:
      tags:
        - Galleries
      summary: Ações em lote (deletar, mudar status, compartilhar)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
                - gallery_ids
              properties:
                action:
                  type: string
                  enum: [delete, change_status, share]
                gallery_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                params:
                  type: object
                  properties:
                    status:
                      type: string
                      enum: [draft, published, client_selection, archived]
      responses:
        '200':
          description: Ação executada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      success_count:
                        type: integer
                      failed_count:
                        type: integer
                      errors:
                        type: array
                        items:
                          type: object
  /galleries/{id}/sync:
    post:
      tags:
        - Galleries
      summary: Sincronizar galeria com Google Drive
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sincronização iniciada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      sync_job_id:
                        type: string
                        format: uuid
                      status:
                        type: string
                        enum: [running]
                      started_at:
                        type: string
                        format: date-time
  /galleries/{id}/sync-status:
    get:
      tags:
        - Galleries
      summary: Verificar status da sincronização
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Status da sincronização
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SyncJob'

  # ==================== PHOTOS ====================
...